Add files via upload
Attached is the the homepage index.html file as well as all of the CSS files.

When inspecting the bottom ul (screen shot attached), you can see that there's a padding-inline-start: 40px; which is shifting the ul content over to the right. However, it says that this is being done on a user agent stylesheet? I have no idea how to access that even after searching on Google, the code is also not in any of the CSS or bootstrap files.